Correção do contexto em Permissoes + Correção da coleção 2 listada em tbcontacorrete + Proteção do privilégio comercial em webpagesroles

+++ api/Negocios/Administracao/GatewayWebpagesRoles.cs
@@ -297,6 +297,7 @@ namespace api.Negocios.Administracao
             // Por segurança, só deixa alterar se o usuário tiver permissão para setar aquela role 
             Int32 RoleLevelMin = Permissoes.GetRoleLevelMin(token);
             if (param.RoleLevel < RoleLevelMin) throw new Exception("401"); // não possui autorização para criar um privilégio com esse RoleLevel
+            if(_db.webpages_Roles.Where(r => r.RoleName.ToUpper().Equals(param.RoleName.ToUpper())).FirstOrDefault() != null) throw new Exception("500"); // já existe um privilégio com esse nome
             _db.webpages_Roles.Add(param);
             _db.SaveChanges();
             return param.RoleId;
@@ -310,6 +311,13 @@ namespace api.Negocios.Administracao
         /// <returns></returns>
         public static void Delete(string token, Int32 RoleId)
         {
+            webpages_Roles role = _db.webpages_Roles.Where(r => r.RoleId == RoleId).FirstOrDefault();
+
+            if(role == null) throw new Exception("500"); // não existe role com o Id informado
+
+            Int32 RoleLevelMin = Permissoes.GetRoleLevelMin(token);
+            if (role.RoleName.ToUpper().Equals("COMERCIAL") || role.RoleLevel < RoleLevelMin) throw new Exception("401"); // não possui autorização para remover o privilégio
+
             GatewayWebpagesPermissions.Delete(token, RoleId);
             GatewayWebpagesUsersInRoles.Delete(token, RoleId, true);
             _db.webpages_Roles.Remove(_db.webpages_Roles.Where(e => e.RoleId.Equals(RoleId)).First());
@@ -330,7 +338,10 @@ namespace api.Negocios.Administracao
                     .First<webpages_Roles>();
 
             if (param.RoleName != null && param.RoleName != value.RoleName)
+            {
+                if(value.RoleName.ToUpper().Equals("COMERCIAL")) throw new Exception("401"); // não possui autorização para alterar o nome privilégio comercial
                 value.RoleName = param.RoleName;
+            }
             if (param.RoleLevel != value.RoleLevel)
             {
                 // Por segurança, só deixa alterar se o usuário tiver permissão para setar aquela role
